use 16-bit integers for mouse motion as per hid report mode

+++ src/platform/amiga/quad_mouse.c
@@ -20,7 +20,7 @@
 #include "hardware/gpio.h"
 
 // mouse motion values, used between core0 and core1
-volatile int8_t x = 0, y = 0;
+volatile int16_t x = 0, y = 0;
 volatile bool motion_flag = false;
 volatile uint8_t motion_divider = 2;
 
@@ -87,7 +87,7 @@ void amiga_quad_mouse_button(enum amiga_quad_mouse_buttons button, bool pressed)
     }
 }
 
-void amiga_quad_mouse_set_motion(int8_t in_x, int8_t in_y)
+void amiga_quad_mouse_set_motion(int16_t in_x, int16_t in_y)
 {
     x = in_x;
     y = in_y;
